requirements.txt
requests==2.10.0
flask==0.11.1
PyJWT==1.4.0
PyMongo>=3.2.0
flask-pymongo==0.4.1
